Check if x-pack already installed

+++ src/entrypoint.sh
@@ -4,7 +4,7 @@ set -m
 
 /run/miscellaneous/restore_config.sh
 /run/miscellaneous/edit_config.sh
-/usr/share/kibana/bin/kibana-plugin install x-pack
+test -f /usr/share/kibana/plugins/x-pack/package.json || /usr/share/kibana/bin/kibana-plugin install x-pack
 /run/miscellaneous/wait_for_elasticsearch.sh
 
 # Run as user "logstash" if the command is "kibana"
